Auth menu styling (#1153)
* Auth menu styling

* Organizing

* Cleaning

* Cleaning

* Cleaning menu
