+++ src/components/TopNav/TopNavLogo/TopNavLogo.scss
@@ -11,7 +11,7 @@
     margin-right: 6px;
   }
 
-  @media (max-width: 992px) {
+  @media (max-width: 1200px) {
     margin-left: 12px;
   }
 }
