+++ src/components/TopNav/TopNavMobileMenu/TopNavMobileMenu.scss
@@ -1,0 +1,152 @@
+$dropdown-height: 316px;
+$mobile-dropdown-height: 1000px;
+$top-nav-height: 60px;
+
+@keyframes addOverlay {
+  from {
+    background: rgba(0, 0, 0, 0);
+  }
+  to {
+    background: rgba(0, 0, 0, 0.3);
+  }
+}
+
+@keyframes mobileSlideDown {
+  from {
+    top: $top-nav-height - $mobile-dropdown-height;
+  }
+  to {
+    top: $top-nav-height;
+  }
+}
+
+@keyframes slideDown {
+  from {
+    top: $top-nav-height - $dropdown-height;
+  }
+  to {
+    top: $top-nav-height;
+  }
+}
+
+.TopNavMobileMenu {
+  display: none;
+
+  @media (max-width: 1200px) {
+    background-color: var(--color-bg);
+    display: block;
+  }
+
+  &__button {
+    background: transparent;
+    border: none;
+    height: $top-nav-height;
+    outline: none;
+    position: relative;
+    transition: background 0.1s ease-in;
+    width: $top-nav-height;
+    z-index: 3;
+
+    &:hover {
+      background: rgba(0, 0, 0, 0.1);
+    }
+  }
+
+  &__chevron-icon {
+    color: var(--color-sail-gray-300);
+    transition: transform 0.3s ease-in-out;
+
+    &--open {
+      transform: rotateX(180deg) translateZ(-1px);
+    }
+  }
+
+  &__column {
+    &:not(:last-child) {
+      margin-right: 60px;
+
+      @media (max-width: 992px) {
+        margin-right: 0;
+      }
+    }
+  }
+
+  &__column-title {
+    background-color: transparent;
+    border: 0;
+    color: var(--color-sail-gray-300);
+    cursor: pointer;
+    margin-bottom: 24px;
+    outline: none;
+    padding: 0;
+  }
+
+  &__dropdown-container {
+    animation: slideDown 0.3s forwards;
+    background: var(--color-white);
+    box-shadow: 0 6px 6px rgba(4, 34, 53, 0.3);
+    display: flex;
+    height: $dropdown-height;
+    left: 0;
+    padding: 60px;
+    position: fixed;
+    top: $top-nav-height;
+    width: 100%;
+    z-index: 2;
+
+    @media (max-width: 992px) {
+      animation: mobileSlideDown 0.3s forwards;
+      flex-direction: column;
+      height: $mobile-dropdown-height;
+      padding: 50px 36px 18px;
+    }
+  }
+
+  &__icon-holder {
+    display: inline-block;
+    position: absolute;
+  }
+
+  &__link {
+    background: var(--color-white);
+    border: none;
+    color: var(--color-primary);
+    display: block;
+    line-height: 1;
+    margin-bottom: 24px;
+    padding: 0;
+    text-align: unset;
+
+    &:hover {
+      text-decoration: underline;
+    }
+
+    @media (max-width: 992px) {
+      margin-bottom: 24px;
+    }
+  }
+
+  &__not-visible-tab {
+    @media (min-width: 992px) {
+      display: none;
+    }
+  }
+
+  &__overlay {
+    animation: addOverlay 0.3s forwards;
+    height: calc(100vh - #{$top-nav-height});
+    left: 0;
+    outline: none;
+    position: fixed;
+    top: $top-nav-height;
+    width: 100%;
+    z-index: 1;
+  }
+
+  &__separator {
+    background-color: #c1c9d2;
+    height: 0.5px;
+    margin-bottom: 20px;
+    width: 40px;
+  }
+}
