+++ src/components/TopNav/TopNavPopoverButton/TopNavPopoverButton.scss
@@ -30,7 +30,7 @@
   &__Popover {
     padding: 18px;
 
-    @media (max-width: 992px) {
+    @media (max-width: 1200px) {
       display: none;
     }
   }
