+++ src/components/TopNav/TopNavText/TopNavText.scss
@@ -21,7 +21,7 @@
   &__Popover {
     padding: 18px;
 
-    @media (max-width: 992px) {
+    @media (max-width: 1200px) {
       display: none;
     }
   }
